backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
pymysql==1.1.0
redis==5.0.1
firebase-admin==6.4.0
pydantic==2.6.1
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.9
